+++ apps/app/pages/settings/plans.tsx
@@ -10,7 +10,6 @@ import { useUser } from "@perfolio/hooks"
 import { withAuthenticationRequired } from "@auth0/auth0-react"
 import { Card } from "@perfolio/ui/components"
 import { CheckCircleIcon } from "@heroicons/react/solid"
-
 type Price = {
   id: string
   value: number
@@ -38,66 +37,79 @@ const ProductCard: React.FC<Product & { selected: "yearly" | "monthly" }> = ({
   features,
   prices,
   selected,
+  name,
+  description,
 }): JSX.Element => {
   const { t } = useI18n()
   const { user, isLoading } = useUser()
   const router = useRouter()
   return (
-    <div className="flex flex-col justify-between space-y-6 lg:flex-row lg:space-y-0">
-      <div className="flex flex-col space-y-8 lg:w-3/5 2xl:w-3/4">
-        <div className="flex items-center">
-          <span className="font-semibold uppercase text-primary whitespace-nowrap">
-            {t("setPlanWhatsIncl")}
-          </span>
-          <div className="flex-grow w-full mx-4 border-b border-primary"></div>
+    <Card>
+      <div className="relative flex flex-col lg:flex-row">
+        <div className="w-full lg:w-3/5 2xl:w-3/4">
+          <Card.Header>
+            <Card.Header.Title title={name} subtitle={description} />
+          </Card.Header>
+          <Card.Content>
+            <div className="flex flex-col justify-between space-y-6 lg:flex-row lg:space-y-0">
+              <div className="flex flex-col w-full space-y-8">
+                <div className="flex items-center">
+                  <span className="font-semibold uppercase text-primary whitespace-nowrap">
+                    {t("setPlanWhatsIncl")}
+                  </span>
+                  <div className="flex-grow w-full mx-4 border-b border-primary"></div>
+                </div>
+                <ul className="grid grid-cols-1 gap-6 lg:grid-cols-2">
+                  {features.map((feature) => (
+                    <li key={feature} className="flex items-center">
+                      <Icon label="Checkmark" size="xs">
+                        <CheckCircleIcon className="text-success" />
+                      </Icon>
+                      <Text>{feature}</Text>
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            </div>
+          </Card.Content>
         </div>
-        <ul className="grid grid-cols-1 gap-6 lg:grid-cols-2">
-          {features.map((feature) => (
-            <li key={feature} className="flex items-center">
-              <Icon label="Checkmark" size="xs">
-                <CheckCircleIcon className="text-success" />
-              </Icon>
-              <Text>{feature}</Text>
-            </li>
-          ))}
-        </ul>
-      </div>
-      <div className="flex flex-col items-center justify-center p-12 -m-12 space-y-10 lg:w-2/5 2xl:w-1/4 bg-gray-50">
-        <div className="flex items-end justify-center">
-          <span className="text-5xl font-bold text-gray-800">{`${
-            prices[selected].value
-          }${getCurrencySymbol(prices[selected].currency)}`}</span>
-          <span className="font-semibold text-gray-700 whitespace-nowrap">
-            / {prices[selected].interval}
-          </span>
+        <div className="flex flex-col items-center justify-center h-full p-12 space-y-10 lg:absolute lg:right-0 lg:inset-y-0 lg:w-2/5 2xl:w-1/4 bg-gray-50">
+          <div className="flex items-end justify-center">
+            <span className="text-5xl font-bold text-gray-800">{`${
+              prices[selected].value
+            }${getCurrencySymbol(prices[selected].currency)}`}</span>
+            <span className="font-semibold text-gray-700 whitespace-nowrap">
+              / {prices[selected].interval}
+            </span>
+          </div>
+          <Button
+            disabled={!user}
+            loading={isLoading}
+            onClick={async () => {
+              if (!user) {
+                console.error(t("setPlanUserError"))
+                return
+              }
+              const res = await fetch(
+                `/api/stripe/checkout?priceId=${prices[selected].id}&customerId=${user.stripeCustomerId}`,
+                {
+                  method: "POST",
+                },
+              )
+              if (res.status !== 200) {
+                console.error(res.body)
+              }
+              const { url } = (await res.json()) as { url: string }
+              router.push(url)
+            }}
+            kind="primary"
+            size="auto"
+          >
+            {t("setPlanSwitchPlanText")}
+          </Button>
         </div>
-        <Button
-          disabled={!user}
-          loading={isLoading}
-          onClick={async () => {
-            if (!user) {
-              console.error(t("setPlanUserError"))
-              return
-            }
-            const res = await fetch(
-              `/api/stripe/checkout/${prices[selected].id}?customerId=${user.stripeCustomerId}`,
-              {
-                method: "POST",
-              },
-            )
-            if (res.status !== 200) {
-              console.error(res.body)
-            }
-            const { url } = (await res.json()) as { url: string }
-            router.push(url)
-          }}
-          kind="primary"
-          size="auto"
-        >
-          {t("setPlanSwitchPlanText")}
-        </Button>
       </div>
-    </div>
+    </Card>
   )
 }
 
@@ -111,10 +123,13 @@ const Page: NextPage<PageProps> = ({ products, translations }) => {
     <AppLayout side="left" sidebar={<SideNavbar />}>
       <div className="flex flex-col space-y-16">
         <Card>
-          <div className="flex flex-col items-center p-8 my-8 space-y-6">
-            <h2 className="text-5xl font-black text-gray-900 ">Pricing Plans</h2>
-
-            <Text size="lg">{t("setPlanPickPlanText")}</Text>
+          <div className="flex flex-col items-center p-8 my-8 space-y-6 text-center">
+            <Card.Header>
+              <Card.Header.Title
+                title="Pricing Plans"
+                subtitle={t("setPlanPickPlanText")}
+              ></Card.Header.Title>
+            </Card.Header>
             <ToggleGroup
               options={[
                 { display: t("setPlanMonthlyBill"), id: "monthly" },
@@ -127,14 +142,7 @@ const Page: NextPage<PageProps> = ({ products, translations }) => {
           </div>
         </Card>
         {products.map((product) => (
-          <Card key={product.id}>
-            <Card.Header>
-              <Card.Header.Title title={product.name} subtitle={product.description} />
-            </Card.Header>
-            <Card.Content>
-              <ProductCard {...product} selected={selected} />
-            </Card.Content>
-          </Card>
+          <ProductCard key={product.id} {...product} selected={selected} />
         ))}
       </div>
     </AppLayout>
